@@ -3,7 +3,7 @@ name = "doccano-client"
 version = "1.2.8"
 description = "A simple client for doccano API."
 authors = [{ name = "Hironsan", email = "hiroki.nakayama.py@gmail.com" }]
-requires-python = "~=3.8"
+requires-python = ">=3.9"
 readme = "README.md"
 license = "MIT"
 classifiers = [
@@ -17,8 +17,8 @@ classifiers = [
 dependencies = [
     "requests>=2.28.1,<3",
     "dataclasses-json>=0.5.7,<0.6",
-    "pydantic>=1.9.2,<2",
-    "requests-toolbelt>=0.9.1,<0.10",
+    "pydantic>=2.0.0",
+    "requests-toolbelt>=1.0.0",
     "pyyaml!=5.4.0,!=5.4.1,!=6.0.0",
 ]
 
@@ -53,7 +53,7 @@ dev = [
     "mkdocs>=1.3.1,<2",
     "mkdocs-material>=8.5.3,<9",
     "mkdocs-same-dir>=0.1.1,<0.2",
-    "vcrpy>=4.2.1,<5",
+    "vcrpy>=5.0.0",
     "mkdocstrings[python, crystal]>=0.19.0,<0.20",
 ]
 
