@@ -1,7 +1,6 @@
-from typing import Dict
+from typing_extensions import Annotated, TypeAlias
 
-from pydantic import BaseModel, root_validator
-from pydantic.types import ConstrainedStr
+from pydantic import BaseModel, StringConstraints, model_validator
 
 
 class UserDetails(BaseModel):
@@ -16,20 +15,24 @@ class PasswordUpdated(BaseModel):
     detail: str
 
 
-class Password(ConstrainedStr):
-    min_length = 2
-    max_length = 128
-    strip_whitespace = True
+PasswordStr: TypeAlias = Annotated[
+    str,
+    StringConstraints(
+        min_length=2,
+        max_length=128,
+        strip_whitespace=True,
+    ),
+]
 
 
 class PasswordChange(BaseModel):
-    new_password: Password
-    confirm_password: Password
+    new_password: PasswordStr
+    confirm_password: PasswordStr
 
-    @root_validator
-    def new_password_matches_confirm_password(cls, values: Dict[str, Password]):
-        new_password = values.get("new_password")
-        confirm_password = values.get("confirm_password")
+    @model_validator(mode="after")
+    def new_password_matches_confirm_password(self) -> "PasswordChange":
+        new_password = self.new_password
+        confirm_password = self.confirm_password
         if new_password != confirm_password:
             raise ValueError("The new password does not match the confirm one.")
-        return values
+        return self
