@@ -24,7 +24,7 @@ class DataDownloadRepository:
         """
         resource = f"projects/{project_id}/download-format"
         response = self._client.get(resource)
-        options = [Option.parse_obj(label) for label in response.json()]
+        options = [Option.model_validate(label) for label in response.json()]
         return options
 
     def find_option_by_name(self, project_id: int, name: str) -> Option:
@@ -46,7 +46,9 @@ class DataDownloadRepository:
                 return option
         raise ValueError(f"Download option '{name}' not found")
 
-    def schedule_download(self, project_id: int, option: Option, only_approved=False) -> str:
+    def schedule_download(
+        self, project_id: int, option: Option, only_approved=False
+    ) -> str:
         """Schedule a download
 
         Args:
@@ -79,7 +81,9 @@ class DataDownloadRepository:
         response = self._client.get(resource, params=params, stream=True)
         content_disposition = response.headers["Content-Disposition"]
         ATTRIBUTE = "filename="
-        file_name = content_disposition[content_disposition.find(ATTRIBUTE) + len(ATTRIBUTE) + 1 : -1]
+        file_name = content_disposition[
+            content_disposition.find(ATTRIBUTE) + len(ATTRIBUTE) + 1 : -1
+        ]
         dir_path = pathlib.Path(dir_name)
         dir_path.mkdir(parents=True, exist_ok=True)
         file_path = dir_path / file_name
