@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import functools
-from typing import Generic, List, TypeVar
+from typing import Generic, List, TypeVar, Type
 
 from doccano_client.models.label import (
     BoundingBox,
@@ -20,7 +20,9 @@ T = TypeVar("T", bound=Label)
 class LabelRepository(Generic[T]):
     """Repository for interacting with the Doccano label API"""
 
-    def __init__(self, client: BaseRepository, label_class: T, resource_type: str):
+    def __init__(
+        self, client: BaseRepository, label_class: Type[T], resource_type: str
+    ):
         self._client = client
         self._label_class = label_class
         self._resource_type = resource_type
@@ -38,7 +40,7 @@ class LabelRepository(Generic[T]):
         """
         resource = f"projects/{project_id}/examples/{example_id}/{self._resource_type}/{label_id}"
         response = self._client.get(resource)
-        return self._label_class.parse_obj(response.json())
+        return self._label_class.model_validate(response.json())
 
     def list(self, project_id: int, example_id: int) -> List[T]:
         """Return all label in which you are a member
@@ -52,7 +54,7 @@ class LabelRepository(Generic[T]):
         """
         resource = f"projects/{project_id}/examples/{example_id}/{self._resource_type}"
         response = self._client.get(resource)
-        labels = [self._label_class.parse_obj(label) for label in response.json()]
+        labels = [self._label_class.model_validate(label) for label in response.json()]
         return labels
 
     def create(self, project_id: int, label: T) -> T:
@@ -65,9 +67,11 @@ class LabelRepository(Generic[T]):
         Returns:
             T: The created label
         """
-        resource = f"projects/{project_id}/examples/{label.example}/{self._resource_type}"
-        response = self._client.post(resource, json=label.dict(exclude={"id"}))
-        return self._label_class.parse_obj(response.json())
+        resource = (
+            f"projects/{project_id}/examples/{label.example}/{self._resource_type}"
+        )
+        response = self._client.post(resource, json=label.model_dump(exclude={"id"}))
+        return self._label_class.model_validate(response.json())
 
     def update(self, project_id: int, label: T) -> T:
         """Update a label
@@ -85,8 +89,8 @@ class LabelRepository(Generic[T]):
         if label.id is None:
             raise ValueError("Label id is required")
         resource = f"projects/{project_id}/examples/{label.example}/{self._resource_type}/{label.id}"
-        response = self._client.put(resource, json=label.dict())
-        return self._label_class.parse_obj(response.json())
+        response = self._client.put(resource, json=label.model_dump())
+        return self._label_class.model_validate(response.json())
 
     def delete(self, project_id: int, label: T):
         """Delete a label
@@ -114,9 +118,21 @@ class LabelRepository(Generic[T]):
         self._client.delete(resource)
 
 
-CategoryRepository = functools.partial(LabelRepository[Category], label_class=Category, resource_type="categories")
-SpanRepository = functools.partial(LabelRepository[Span], label_class=Span, resource_type="spans")
-RelationRepository = functools.partial(LabelRepository[Relation], label_class=Relation, resource_type="relations")
-SegmentRepository = functools.partial(LabelRepository[Segment], label_class=Segment, resource_type="segments")
-TextRepository = functools.partial(LabelRepository[Text], label_class=Text, resource_type="texts")
-BoundingBoxRepository = functools.partial(LabelRepository[BoundingBox], label_class=BoundingBox, resource_type="bboxes")
+CategoryRepository = functools.partial(
+    LabelRepository[Category], label_class=Category, resource_type="categories"
+)
+SpanRepository = functools.partial(
+    LabelRepository[Span], label_class=Span, resource_type="spans"
+)
+RelationRepository = functools.partial(
+    LabelRepository[Relation], label_class=Relation, resource_type="relations"
+)
+SegmentRepository = functools.partial(
+    LabelRepository[Segment], label_class=Segment, resource_type="segments"
+)
+TextRepository = functools.partial(
+    LabelRepository[Text], label_class=Text, resource_type="texts"
+)
+BoundingBoxRepository = functools.partial(
+    LabelRepository[BoundingBox], label_class=BoundingBox, resource_type="bboxes"
+)
