@@ -4,9 +4,9 @@ from pydantic import BaseModel
 
 
 class Comment(BaseModel):
-    id: Optional[int]
+    id: Optional[int] = None
     text: str = ""
     example: int
-    user: Optional[int]
-    username: Optional[str]
-    created_at: Optional[str]
+    user: Optional[int] = None
+    username: Optional[str] = None
+    created_at: Optional[str] = None
